requirements.txt
fastapi
uvicorn
python-multipart
aiofiles
numpy
opencv-python-headless
pillow
rembg[cli]
torch
torchvision
transformers
scikit-image
jinja2